chore: improvements

+++ src/utils/replaceInFile.ts
@@ -1,6 +1,10 @@
 import fs from 'fs-extra'
 
-export const replaceInFile = (path: string, searchValue: string, replaceValue: string) => {
+export const replaceInFile = (path: string, searchValue: string | RegExp, replaceValue: string) => {
+    if (!fs.existsSync(path)) {
+        return
+    }
+
     const file = fs.readFileSync(path, 'utf-8')
     fs.rmSync(path)
     fs.writeFileSync(path, file.replaceAll(searchValue, replaceValue))
